+++ subsys/bluetooth/host/id.c
@@ -13,6 +13,7 @@
 
 #include "hci_core.h"
 #include "id.h"
+#include "scan.h"
 #include "smp.h"
 #include "conn_internal.h"
 #include "keys.h"
